Refactor hooks imports
