+++ src/components/general/CodeBlock.tsx
@@ -2,7 +2,7 @@ import { ComponentProps, forwardRef, useCallback, useState } from 'react';
 import { Highlight } from 'prism-react-renderer';
 
 import { IconButton } from 'components/general';
-import useDarkTheme from 'hooks/useDarkTheme';
+import { useDarkTheme } from 'hooks';
 import { CopyIcon, DoneIcon } from 'assets/icons';
 import cn from 'utils/cn';
 import { prismThemes } from 'constants';
